refactor: move BarcodeScanner in `src/feature/idpay/barcode`
