+++ ts/features/idpay/payment/components/Barcode/__tests__/BarcodeScanner.test.tsx
@@ -1,6 +1,6 @@
 import * as O from "fp-ts/lib/Option";
 import { Barcode, BarcodeFormat } from "vision-camera-code-scanner";
-import { getIOBarcodeType, retrieveNextBarcode } from "../BarcodeScanner";
+import { retrieveNextBarcode } from "../BarcodeScanner";
 
 describe("test retrieveNextBarcode function", () => {
   it("should return `O.none` because of an empty array as input", () => {
@@ -52,29 +52,5 @@ describe("test retrieveNextBarcode function", () => {
     const output = O.toUndefined(retrieveNextBarcode(input));
 
     expect(output?.format).toBe("DATA_MATRIX");
-  });
-});
-
-describe("test getIOBarcodeType function", () => {
-  it("should return unknown if empty value", () => {
-    const input = "";
-    const output = getIOBarcodeType(input);
-
-    expect(output).toBe("UNKNOWN");
-  });
-
-  describe("test IDPAY barcode type", () => {
-    it("should return IDPAY on valid QRCode content", () => {
-      const input = "https://continua.io.pagopa.it/idpay/auth/mkdb1yxg";
-      const output = getIOBarcodeType(input);
-
-      expect(output).toBe("IDPAY");
-    });
-    it("should return UNKNOWN on invalid QR content", () => {
-      const input = "https://continua.io.pagopa.it/idpay/auth/pippo";
-      const output = getIOBarcodeType(input);
-
-      expect(output).toBe("UNKNOWN");
-    });
   });
 });
