+++ ts/features/idpay/payment/screens/__tests__/IDPayPaymentCodeScanScreen.test.tsx
@@ -17,7 +17,7 @@ jest.mock("../../components/BarcodeScanner", () => ({
 import {
   IOBarcodeScanner,
   useIOBarcodeScanner
-} from "../../components/BarcodeScanner";
+} from "../../components/Barcode/BarcodeScanner";
 
 const mockCameraComponent = <View testID="cameraComponentTestID" />;
